+++ pom.xml
@@ -308,6 +308,7 @@
         <module>storm-buildtools/maven-shade-clojure-transformer</module>
         <module>storm-buildtools/storm-maven-plugins</module>
         <module>storm-core</module>
+        <module>storm-drpc-server</module>
         <module>storm-rename-hack</module>
         <module>storm-clojure</module>
