+++ storm-core/pom.xml
@@ -372,18 +372,6 @@
                 </exclusion>
             </exclusions>
         </dependency>
-        <dependency>
-           <groupId>org.glassfish.jersey.core</groupId>
-            <artifactId>jersey-server</artifactId>
-        </dependency> 
-        <dependency>
-            <groupId>org.glassfish.jersey.containers</groupId> 
-            <artifactId>jersey-container-servlet-core</artifactId>
-        </dependency> 
-        <dependency>
-            <groupId>org.glassfish.jersey.containers</groupId> 
-            <artifactId>jersey-container-jetty-http</artifactId>
-        </dependency>
     </dependencies>
     <build>
         <sourceDirectory>src/jvm</sourceDirectory>
