+++ storm-drpc-server/src/main/java/org/apache/storm/daemon/drpc/webapp/DRPCResource.java
@@ -26,9 +26,9 @@ import javax.ws.rs.core.Context;
 
 import org.apache.storm.daemon.drpc.DRPC;
 import org.apache.storm.metric.StormMetricsRegistry;
-import org.apache.thrift.TException;
+import org.apache.storm.thrift.TException;
 
-import com.codahale.metrics.Meter;
+import org.apache.storm.shade.com.codahale.metrics.Meter;
 
 @Path("/drpc/")
 public class DRPCResource {
